+++ changepoint/src/bocpd.rs
@@ -14,7 +14,7 @@ use serde::{Deserialize, Serialize};
 
 /// Online Bayesian Change Point Detection state container.
 #[cfg_attr(feature = "serde1", derive(Serialize, Deserialize))]
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct Bocpd<X, Fx, Pr>
 where
     Fx: Rv<X> + HasSuffStat<X>,