+++ pychangepoint/src/bocpd.rs
@@ -1,15 +1,20 @@
 use crate::convert;
 use changepoint::BocpdLike;
 use nalgebra::DVector;
-use pyo3::exceptions::PyValueError;
+use pyo3::exceptions::{PyTypeError, PyValueError};
 use pyo3::prelude::*;
+use pyo3::pyclass::CompareOp;
+use pyo3::types::PyTuple;
 use rv::dist::{
     Bernoulli, Beta, Gamma, Gaussian, MvGaussian, NormalGamma,
     NormalInvChiSquared, NormalInvGamma, NormalInvWishart, Poisson,
 };
 
+use bincode::{deserialize, serialize};
+use serde::{Deserialize, Serialize};
+
 /// The variant of the prior distribution
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum PriorVariant {
     NormalGamma(NormalGamma),
     NormalInvGamma(NormalInvGamma),
@@ -21,14 +26,61 @@ pub enum PriorVariant {
 
 /// Prior distribution, which also describes the liklihood distribution of the
 /// change point detector.
-#[pyclass]
-#[derive(Clone, Debug)]
+#[pyclass(module = "changepoint")]
+#[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Prior {
     pub dist: PriorVariant,
 }
 
+macro_rules! count {
+    () => (0usize);
+    ( $x:tt $($xs:tt)* ) => (1usize + count!($($xs)*));
+}
+
+macro_rules! handle_kind {
+    ($name: ident, $args: ident, $($idx:tt $n:tt),+) => {{
+        if $args.len() == count!($($idx)*) {
+            Self::$name(
+                $(
+                    $args.get_item($idx)?.extract()?,
+                )+
+            )
+        } else {
+            Err(PyTypeError::new_err(format!("Prior kind '{}' requires the following arguments: {}", stringify!($name), stringify!($($n),+))))
+        }
+    }}
+}
+
 #[pymethods]
 impl Prior {
+    #[new]
+    #[pyo3(signature = (kind, *args))]
+    pub fn new(kind: &str, args: &PyTuple) -> PyResult<Self> {
+        match kind {
+            "normal_gamma" => {
+                handle_kind!(normal_gamma, args, 0 m, 1 r, 2 s, 3 v)
+            }
+            "normal_inv_gamma" => {
+                handle_kind!(normal_inv_gamma, args, 0 m, 1 v, 2 a, 3 b)
+            }
+            "normal_inv_chi_squared" => {
+                handle_kind!(normal_inv_chi_squared, args, 0 m, 1 k, 2 v, 3 s2)
+            }
+            "normal_inv_wishart" => {
+                handle_kind!(normal_inv_wishart, args, 0 m, 1 k, 2 df, 3 scale)
+            }
+            "beta_bernoulli" => {
+                handle_kind!(beta_bernoulli, args, 0 alpha, 1 beta)
+            }
+            "poisson_gamma" => {
+                handle_kind!(poisson_gamma, args, 0 shape, 1 rate)
+            }
+            unknown_kind => Err(PyTypeError::new_err(format!(
+                "Unknown prior kind '{unknown_kind}'"
+            ))),
+        }
+    }
+
     #[staticmethod]
     #[pyo3(signature = (m = 0.0, r = 1.0, s = 1.0, v = 1.0))]
     pub fn normal_gamma(m: f64, r: f64, s: f64, v: f64) -> PyResult<Self> {
@@ -142,6 +194,19 @@ impl Prior {
                 g.rate()
             ),
         }
+    }
+
+    pub fn __setstate__(&mut self, state: Vec<u8>) -> PyResult<()> {
+        self.dist = deserialize(&state).unwrap();
+        Ok(())
+    }
+
+    pub fn __getstate__(&self) -> PyResult<Vec<u8>> {
+        Ok(serialize(&self.dist).unwrap())
+    }
+
+    pub fn __getnewargs__(&self) -> PyResult<(String, f64, f64, f64, f64)> {
+        Ok(("normal_gamma".to_string(), 0.0, 1.0, 1.0, 1.0))
     }
 }
 
@@ -321,7 +386,7 @@ fn dist_to_bocpd(dist: Prior, lambda: f64) -> BocpdVariant {
 
 /// The variant of the `Bocpd`. Describes the prior, likelihood, and the input
 /// data type.
-#[derive(Clone, Debug)]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
 pub enum BocpdVariant {
     NormalGamma(changepoint::Bocpd<f64, Gaussian, NormalGamma>),
     NormalInvGamma(changepoint::Bocpd<f64, Gaussian, NormalInvGamma>),
@@ -379,8 +444,8 @@ impl BocpdVariant {
 }
 
 /// Online Bayesian Change Point Detection state container
-#[derive(Clone, Debug)]
-#[pyclass]
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+#[pyclass(module = "changepoint")]
 /// Create a new BOCPD
 ///
 /// Parameters
@@ -422,5 +487,29 @@ impl Bocpd {
     /// Observe a new datum. Returns the run length probabilities for each step.
     pub fn step(&mut self, datum: &PyAny) -> PyResult<Vec<f64>> {
         self.bocpd.step(datum).map(|rs| rs.to_vec())
+    }
+
+    pub fn __setstate__(&mut self, state: Vec<u8>) -> PyResult<()> {
+        self.bocpd = deserialize(&state).unwrap();
+        Ok(())
+    }
+
+    pub fn __getstate__(&self) -> PyResult<Vec<u8>> {
+        Ok(serialize(&self.bocpd).unwrap())
+    }
+
+    pub fn __getnewargs__(&self) -> PyResult<(Prior, f64)> {
+        Ok((Prior::beta_bernoulli(0.5, 0.5).unwrap(), 1.0))
+    }
+
+    fn __richcmp__(&self, other: &Self, op: CompareOp) -> PyResult<bool> {
+        match op {
+            CompareOp::Lt => Ok(false),
+            CompareOp::Le => Ok(false),
+            CompareOp::Eq => Ok(self.bocpd == other.bocpd),
+            CompareOp::Ne => Ok(self.bocpd != other.bocpd),
+            CompareOp::Gt => Ok(false),
+            CompareOp::Ge => Ok(false),
+        }
     }
 }